First scan - then process
In order to get a better impression for the progress of the operation
we now first scan all directories and push all work that needs to be
done into another process channel which is emptied once scanning is
complete. We use the same parallelization for scanning and processing.

And we do not use the DirEntry objects anymore because they keep file
descriptors open and now when keeping all the flavour objects with all
the descriptors we may likely run out of file descriptors. And there
is no real benefit from keeping the DirEntry objects because we
operate on the filename and path only. So just keep these.